Transaction xdr (#6)
Fixing XDR & ability to accept transaction w/ max_fee

+++ txnbuild/helpers_test.go
@@ -107,7 +107,7 @@ func convertToV0(tx *Transaction) {
 	tx.envelope.V0 = &xdr.TransactionV0Envelope{
 		Tx: xdr.TransactionV0{
 			SourceAccountEd25519: *tx.envelope.SourceAccount().Ed25519,
-			Fee:                  xdr.Uint32(tx.envelope.Fee()),
+			Fee:                  xdr.Uint64(tx.envelope.Fee()),
 			SeqNum:               xdr.SequenceNumber(tx.envelope.SeqNum()),
 			TimeBounds:           tx.envelope.V1.Tx.TimeBounds,
 			Memo:                 tx.envelope.Memo(),
